Move layouts into folders for testing and fix redirects
