+++ src/layout/App/App.tsx
@@ -1,4 +1,4 @@
-import SideBar from "./Sidebar";
+import SideBar from "../Sidebar/Sidebar";
 import styles from "./App.module.css";
 import { Outlet } from "react-router";
